refactor(destwebhook): scope proxy transport to webhook package
Move proxytransport.go, the Envoy flag mapping, and all sentinels
(ErrProxyInfra, ErrProxyDestination, IsProxyInfraError) into the
destwebhook package — the only place they're actually used. destregistry
keeps a generic WrapTransport hook on HTTPClientConfig; destwebhook
exports a WrapTransport function and the two webhook providers
(destwebhook, destwebhookstandard) plug it in when constructing their
HTTP client.

destregistry/ no longer has any proxy-specific code. The wrapper, the
Envoy flag table, the CONNECT-response callback, and all four test files
move with the sentinels.

No behavior change.
