+++ internal/destregistry/providers/destwebhook/proxytransport_internal_test.go
@@ -1,4 +1,4 @@
-package destregistry
+package destwebhook
 
 import (
 	"io"
