+++ internal/destregistry/providers/destwebhook/proxytransport_pin_test.go
@@ -1,4 +1,4 @@
-package destregistry_test
+package destwebhook_test
 
 import (
 	"net/http"
@@ -7,7 +7,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/hookdeck/outpost/internal/destregistry"
+	"github.com/hookdeck/outpost/internal/destregistry/providers/destwebhook"
 	"github.com/stretchr/testify/require"
 )
 
@@ -57,6 +57,6 @@ func TestProxyTransport_PinProxyconnectWording(t *testing.T) {
 	wrappedClient := makeProxiedClient(t, unreachable)
 	_, err = wrappedClient.Get("https://example.invalid/")
 	require.Error(t, err)
-	require.True(t, destregistry.IsProxyInfraError(err),
+	require.True(t, destwebhook.IsProxyInfraError(err),
 		"wrapper failed to identify proxy-unreachable as infra; err = %v", err)
 }
